[Refactor] Move snapshot to data_workflows
