+++ be/src/data_workflows/snapshot/snapshot_loader.cpp
@@ -12,27 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/be/src/runtime/snapshot_loader.cpp
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-#include "runtime/snapshot_loader.h"
+#include "data_workflows/snapshot/snapshot_loader.h"
 
 #include <cstdint>
 #include <filesystem>
@@ -43,7 +23,6 @@
 #include "common/logging.h"
 #include "common/system/master_info.h"
 #include "fs/fs.h"
-#include "fs/fs_broker.h"
 #include "fs/fs_factory.h"
 #include "fs/fs_util.h"
 #include "gen_cpp/FileBrokerService_types.h"
@@ -52,6 +31,7 @@
 #include "gen_cpp/HeartbeatService_types.h"
 #include "gen_cpp/TFileBrokerService.h"
 #include "platform/broker_mgr.h"
+#include "platform/fs_broker.h"
 #include "runtime/exec_env.h"
 #include "storage/index/index_descriptor.h"
 #ifndef __APPLE__
