+++ be/src/exec/data_sinks/file_result_writer.cpp
@@ -45,9 +45,9 @@
 #include "exec/plain_text_builder.h"
 #include "formats/csv/converter.h"
 #include "formats/io/formatted_output_stream.h"
-#include "fs/fs_broker.h"
 #include "fs/fs_factory.h"
 #include "gutil/strings/substitute.h"
+#include "platform/fs_broker.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks {
