+++ be/src/exec/file_scanner/parquet_scanner.cpp
@@ -23,7 +23,6 @@
 #include "compute_env/load_path/rejected_record_writer.h"
 #include "exprs/cast_expr.h"
 #include "exprs/column_ref.h"
-#include "fs/fs_broker.h"
 #include "rapidjson/document.h"
 #include "rapidjson/stringbuffer.h"
 #include "rapidjson/writer.h"
