+++ be/src/exec/pipeline/sink/export_sink_operator.cpp
@@ -28,8 +28,8 @@
 #include "exprs/expr_factory.h"
 #include "formats/csv/converter.h"
 #include "formats/io/formatted_output_stream.h"
-#include "fs/fs_broker.h"
 #include "fs/fs_factory.h"
+#include "platform/fs_broker.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks::pipeline {
