+++ be/src/platform/CMakeLists.txt
@@ -22,6 +22,7 @@ set(PLATFORM_FILES
   aws/poco_common.cpp
   aws/poco_http_client.cpp
   aws/poco_http_client_factory.cpp
+  fs_broker.cpp
   broker_mgr.cpp
   download_util.cpp
   key_cache.cpp
