+++ be/src/platform/fs_broker.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "fs/fs_broker.h"
+#include "platform/fs_broker.h"
 
 #include <fmt/format.h>
