+++ be/src/service/service_be/lake_service.cpp
@@ -35,13 +35,13 @@
 #include "common/thread/thread.h"
 #include "common/thread/threadpool.h"
 #include "data_workflows/load/tablet_writer/load_channel_mgr.h"
+#include "data_workflows/snapshot/lake_snapshot_loader.h"
 #include "exec/write_combined_txn_log.h"
 #include "fs/fs_factory.h"
 #include "fs/fs_util.h"
 #include "gen_cpp/tablet_schema.pb.h"
 #include "gutil/strings/join.h"
 #include "runtime/exec_env.h"
-#include "runtime/lake_snapshot_loader.h"
 #include "storage/lake/compaction_policy.h"
 #include "storage/lake/compaction_scheduler.h"
 #include "storage/lake/compaction_task.h"
